backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg[binary]==3.2.10
python-dotenv==1.0.1
